Release/v0.11.0

+++ devtest/test_game/world_state/equipment_info.hpp
@@ -22,7 +22,7 @@ struct EquipSlot {
     }
 
     void unequip() {
-        equipped_id_ = IdT::None;
+        equipped_id_ = static_cast<uint32_t>(IdT::None);
         equipped_ = false;
     }
